lesson 2.1 testdata
