+++ addressbook-web-test/AccountData.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace addressbook_web_test
 {
-    internal class AccountData
+    public class AccountData
     {
         private string username;
         private string password;
