+++ addressbook-web-test/ContactData.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace addressbook_web_test
 {
-    internal class ContactData
+    public class ContactData
     {
         private string firstname;
